Move error handling to dedicated files and update CMakeLists.txt
