+++ include/error_handling.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <portaudio.h>
+
+#include <string>
+
+void LogError(const std::string& context, const std::string& message);
+
+bool Mpg123Succeeded(const std::string& context, int error);
+bool PortAudioSucceeded(const std::string& context, PaError error);
+bool Succeeded(const std::string& context, bool error);
